+++ include/tracey/constants.hpp
@@ -6,7 +6,7 @@ namespace constants {
 const double pi = 3.14159265358979323846;
 }
 
-inline float clamp(float x, float min, float max) {
+inline double clamp(double x, double min, double max) {
     if (x < min) return min;
     if (x > max) return max;
     return x;
