+++ include/tracey/ray/hittable.hpp
@@ -10,13 +10,13 @@ namespace Tracey{
 struct HitRecord
 {
 	//T t;
-	float t;
-	Vector3f point;
-	Vector3f normal;
+	double t;
+	Vector3d point;
+	Vector3d normal;
 	//Material* material;
 	bool front_face;
 
-	inline void set_face_normal(const Tracey::Ray& ray, const Vector3f& outward_normal) {
+	inline void set_face_normal(const Tracey::Ray& ray, const Vector3d& outward_normal) {
 		front_face = ray.get_direction().dot(outward_normal) < 0;
         normal = front_face ? outward_normal :-outward_normal;
     }
@@ -24,7 +24,7 @@ struct HitRecord
 
 class Hittable {
     public:
-        virtual bool Hit(const Tracey::Ray& r, float min, float max, HitRecord& rec) const = 0;
+        virtual bool Hit(const Tracey::Ray& r, double min, double max, HitRecord& rec) const = 0;
 };
 } //namespace Tracey
